+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Fixes
 
 - try and catch blocks when dealing with `fetchUser()`, so that it works when server is down
+- Corrects application name for Android
 
 # 1.2.0 (**8/10/2017**)